fix: auto-grading

+++ be/src/main/java/com/assignment/demo/repository/AssignmentSectionRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AssignmentSectionRepository extends JpaRepository<AssignmentSection,Integer> {
+public interface AssignmentSectionRepository extends JpaRepository<AssignmentSection,Long> {
     List<AssignmentSection> findByAssignmentId(long id);
 }
